pyproject.toml
[project]
name = "youtube-rag"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "langchain-openai>=0.1.14",
    "azure-identity>=1.17.1",
    "python-dotenv>=1.0.1",
    "langchain>=0.2.6",
    "whisper>=1.1.10",
    "pytube>=15.0.0",
    "openai-whisper>=20231117",
    "ffmpeg>=1.4",
    "langchain-community>=0.2.6",
    "scikit-learn>=1.5.1",
    "docarray>=0.40.0",
    "pytubefix>=6.5.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rag"]
